+++ R/ragg-package.R
@@ -1,6 +1,7 @@
 #' @keywords internal
 #' @useDynLib ragg, .registration = TRUE
 #' @importFrom systemfonts match_font
+#' @importFrom textshaping string_width
 "_PACKAGE"
 
 # The following block is used by usethis to automatically manage